Devel

+++ idpy/IdpyCode/IdpySims.py
@@ -34,9 +34,17 @@ class IdpySims(threading.Thread):
         threading.Thread.__init__(self)
         self.sims_vars = {}
         self.sims_idpy_memory = {}
-        self.sims_dump_vars = []
-        self.sims_dump_idpy_memory = []
         self.daemon_flag = daemon_flag
+        '''
+        sims_dump_vars, sims_dump_idpy_memory:
+        lists contaiing the dictionary key to be dumped
+        if empty all key are dumped: these are to be set in the
+        child class: likely specified in a child class of a general
+        class
+        '''
+        self.sims_dump_vars, self.sims_dump_idpy_memory = [], []
+        self.sims_dump_vars_flag, self.sims_dump_idpy_memory_flag = True, True
+
 
     def DumpSnapshot(self, file_name = None, custom_types = None):
         if file_name is None:
@@ -49,27 +57,29 @@ class IdpySims(threading.Thread):
         _grp  = _out_f.create_group(self.__class__.__name__)
 
         '''
+        dumping vars
+        '''
+        if self.sims_dump_vars_flag:
+            _grp_vars = _grp.create_group("vars")
+
+            for key in self.sims_vars:
+                if len(self.sims_dump_vars) == 0 or key in self.sims_dump_vars:
+                    _grp_vars.create_dataset(key, data = self.sims_vars[key])
+        
+        '''
         dumping idpy_memory
         '''
-        _grp_vars = _grp.create_group("vars")
-        
-        for key in self.sims_vars:
-            if len(self.sims_dump_vars) == 0 or key in self.sims_dump_vars:
-                _grp_vars.create_dataset(key, data = self.sims_vars[key])
-        
-        '''
-        dumping idpy_memory
-        '''
-        _grp_idpy_memory = _grp.create_group("idpy_memory")
-        
-        for key in self.sims_idpy_memory:
-            if self.sims_idpy_memory[key] is None:
-                _out_f.close()
-                os.remove(file_name)
-                raise Exception("Cannot dump ", key, ": not allocated")
-            else:
-                if len(self.sims_idpy_memory) == 0 or key in self.sims_dump_idpy_memory:
-                    _grp_idpy_memory.create_dataset(key, data = self.sims_idpy_memory[key].D2H())
+        if self.sims_dump_idpy_memory_flag:
+            _grp_idpy_memory = _grp.create_group("idpy_memory")
+
+            for key in self.sims_idpy_memory:
+                if self.sims_idpy_memory[key] is None:
+                    _out_f.close()
+                    os.remove(file_name)
+                    raise Exception("Cannot dump ", key, ": not allocated")
+                else:
+                    if len(self.sims_dump_idpy_memory) == 0 or key in self.sims_dump_idpy_memory:
+                        _grp_idpy_memory.create_dataset(key, data = self.sims_idpy_memory[key].D2H())
 
         '''
         dumping custom_types
@@ -81,3 +91,29 @@ class IdpySims(threading.Thread):
             
         _out_f.close()
 
+    def ReadSnapshotData(self, file_name = None, full_key = None):
+        if file_name is None:
+            raise Exception("Parameter file_name must not be None")
+        if full_key is None:
+            raise Exception("Parameter full_key must not be None")
+
+        _in_f = h5py.File(file_name, "r")
+        _sims_class_name = list(_in_f.keys())[0]
+        if  _sims_class_name != self.__class__.__name__:
+            print("File class: ", _sims_class_name)
+            print("Present class: ", self.__class__.__name__)
+            raise Exception("The file you are reading has been created by another class!")
+
+        if not full_key in _in_f:
+            raise Exception("Key", full_key, "cannot be found in", file_name)
+        _swap_data = np.array(_in_f.get(full_key))
+        _in_f.close()
+        return _swap_data
+        
+
+    def ReadSnapshotFull(self):
+        '''
+        ReadSnapshotFull:
+        the idea is to use the ReadSnapshotData applied to all cases
+        '''
+        pass
